functions/requirements.txt
requests>=2.32.0
google-cloud-logging>=3.10.0
